backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dnspython==2.6.1
email_validator==2.1.2
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
makefun==1.15.6
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pwdlib==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
rich==13.8.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomli==2.0.1
typer==0.12.5
typing_extensions==4.12.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.1
